+++ src/main/java/racingcar/application/implement/CarRaceHistoryWriter.java
@@ -1,6 +1,6 @@
 package racingcar.application.implement;
 
-import static racingcar.common.FormatConstants.LINE_BREAKER;
+import static racingcar.common.constant.FormatConstants.LINE_BREAKER;
 
 import java.util.List;
 import java.util.stream.Collectors;
