+++ src/main/java/racingcar/application/implement/CarRaceStarter.java
@@ -1,8 +1,8 @@
 package racingcar.application.implement;
 
-import static racingcar.common.RacingCarConstant.MAX_RANDOM_BOUND;
-import static racingcar.common.RacingCarConstant.MIN_RANDOM_BOUND;
-import static racingcar.common.RacingCarConstant.MOVE_THRESHOLD;
+import static racingcar.common.constant.RacingCarConstant.MAX_RANDOM_BOUND;
+import static racingcar.common.constant.RacingCarConstant.MIN_RANDOM_BOUND;
+import static racingcar.common.constant.RacingCarConstant.MOVE_THRESHOLD;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
