+++ src/main/java/racingcar/common/constant/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package racingcar.common;
+package racingcar.common.constant;
 
 public final class ExceptionMessage {
